+++ app/src/main/java/com/cs4530/lifestyleapp/WeatherData.kt
@@ -4,10 +4,10 @@ import com.google.gson.annotations.SerializedName
 
 class WeatherData {
     @SerializedName("temp")
-    var temperature: Double? = null
+    var temperature: Int? = null
     @SerializedName("temp_max")
-    var tempHigh: Double? = null
+    var tempHigh: Int? = null
     @SerializedName("temp_min")
-    var tempLow: Double? = null
+    var tempLow: Int? = null
     var humidity: Double? = null
 }
